REFACTOR: Moved messages and tests of blob module to own file.
