+++ pkg/indexer/decode/handle/blob.go
@@ -1,0 +1,40 @@
+package handle
+
+import (
+	"github.com/celestiaorg/celestia-app/pkg/namespace"
+	appBlobTypes "github.com/celestiaorg/celestia-app/x/blob/types"
+	"github.com/dipdup-io/celestia-indexer/internal/storage"
+	storageTypes "github.com/dipdup-io/celestia-indexer/internal/storage/types"
+	"github.com/dipdup-io/celestia-indexer/pkg/types"
+	"github.com/pkg/errors"
+)
+
+func MsgPayForBlobs(level types.Level, m *appBlobTypes.MsgPayForBlobs) (storageTypes.MsgType, []storage.AddressWithType, []storage.Namespace, uint64, error) {
+	var blobsSize uint64
+	namespaces := make([]storage.Namespace, len(m.Namespaces))
+
+	for nsI, ns := range m.Namespaces {
+		if len(m.BlobSizes) < nsI {
+			return storageTypes.MsgUnknown, nil, nil, 0, errors.Errorf(
+				"blob sizes length=%d is less then namespaces index=%d", len(m.BlobSizes), nsI)
+		}
+
+		appNS := namespace.Namespace{Version: ns[0], ID: ns[1:]}
+		size := uint64(m.BlobSizes[nsI])
+		blobsSize += size
+		namespaces[nsI] = storage.Namespace{
+			FirstHeight: level,
+			Version:     appNS.Version,
+			NamespaceID: appNS.ID,
+			Size:        size,
+			PfbCount:    1,
+			Reserved:    appNS.IsReserved(),
+		}
+	}
+
+	addresses, err := createAddresses(addressesData{
+		{t: storageTypes.MsgAddressTypeSigner, address: m.Signer},
+	}, level)
+
+	return storageTypes.MsgPayForBlobs, addresses, namespaces, blobsSize, err
+}
