+++ pkg/indexer/decode/handle/handlers.go
@@ -1,8 +1,6 @@
 package handle
 
 import (
-	"github.com/celestiaorg/celestia-app/pkg/namespace"
-	types8 "github.com/celestiaorg/celestia-app/x/blob/types"
 	types9 "github.com/celestiaorg/celestia-app/x/qgb/types"
 	types7 "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 	types6 "github.com/cosmos/cosmos-sdk/x/bank/types"
@@ -12,7 +10,6 @@ import (
 	"github.com/dipdup-io/celestia-indexer/internal/storage"
 	"github.com/dipdup-io/celestia-indexer/internal/storage/types"
 	types2 "github.com/dipdup-io/celestia-indexer/pkg/types"
-	"github.com/pkg/errors"
 	"github.com/shopspring/decimal"
 )
 
@@ -151,36 +148,6 @@ func MsgCreatePeriodicVestingAccount(level types2.Level, m *types7.MsgCreatePeri
 		{t: types.MsgAddressTypeToAddress, address: m.ToAddress},
 	}, level)
 	return msgType, addresses, err
-}
-
-func MsgPayForBlobs(level types2.Level, m *types8.MsgPayForBlobs) (types.MsgType, []storage.AddressWithType, []storage.Namespace, uint64, error) {
-	var blobsSize uint64
-	namespaces := make([]storage.Namespace, len(m.Namespaces))
-
-	for nsI, ns := range m.Namespaces {
-		if len(m.BlobSizes) < nsI {
-			return types.MsgUnknown, nil, nil, 0, errors.Errorf(
-				"blob sizes length=%d is less then namespaces index=%d", len(m.BlobSizes), nsI)
-		}
-
-		appNS := namespace.Namespace{Version: ns[0], ID: ns[1:]}
-		size := uint64(m.BlobSizes[nsI])
-		blobsSize += size
-		namespaces[nsI] = storage.Namespace{
-			FirstHeight: level,
-			Version:     appNS.Version,
-			NamespaceID: appNS.ID,
-			Size:        size,
-			PfbCount:    1,
-			Reserved:    appNS.IsReserved(),
-		}
-	}
-
-	addresses, err := createAddresses(addressesData{
-		{t: types.MsgAddressTypeSigner, address: m.Signer},
-	}, level)
-
-	return types.MsgPayForBlobs, addresses, namespaces, blobsSize, err
 }
 
 func MsgGrantAllowance(level types2.Level, m *feegrant.MsgGrantAllowance) (types.MsgType, []storage.AddressWithType, error) {
