+++ pkg/indexer/decode/handle/test/blob_test.go
@@ -1,0 +1,79 @@
+package handle_test
+
+import (
+	appBlobTypes "github.com/celestiaorg/celestia-app/x/blob/types"
+	"github.com/cosmos/cosmos-sdk/types"
+	"github.com/dipdup-io/celestia-indexer/internal/storage"
+	storageTypes "github.com/dipdup-io/celestia-indexer/internal/storage/types"
+	"github.com/dipdup-io/celestia-indexer/internal/test_suite"
+	"github.com/dipdup-io/celestia-indexer/pkg/indexer/decode"
+	"github.com/fatih/structs"
+	"github.com/shopspring/decimal"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+// MsgPayForBlob
+
+func createMsgPayForBlob() types.Msg {
+	msgPayForBlob := appBlobTypes.MsgPayForBlobs{
+		Signer:           "celestia1zefjxuq43xmjq9x4hhw23wkvvz6st5uhv40tys",
+		Namespaces:       [][]byte{{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 189, 44, 204, 197, 144, 206, 197, 121, 37, 22}},
+		BlobSizes:        []uint32{1},
+		ShareCommitments: [][]byte{{176, 28, 134, 119, 32, 117, 87, 107, 231, 67, 121, 255, 209, 106, 52, 99, 88, 183, 85, 36, 67, 137, 98, 199, 144, 159, 13, 178, 111, 190, 121, 36}},
+		ShareVersions:    []uint32{0},
+	}
+	return &msgPayForBlob
+}
+
+func TestDecodeMsg_SuccessOnPayForBlob(t *testing.T) {
+	msgPayForBlob := createMsgPayForBlob()
+	blob, now := testsuite.EmptyBlock()
+	position := 0
+
+	dm, err := decode.Message(msgPayForBlob, blob.Height, blob.Block.Time, position, storageTypes.StatusSuccess)
+
+	addressesExpected := []storage.AddressWithType{
+		{
+			Type: storageTypes.MsgAddressTypeSigner,
+			Address: storage.Address{
+				Id:         0,
+				Height:     blob.Height,
+				LastHeight: blob.Height,
+				Address:    "celestia1zefjxuq43xmjq9x4hhw23wkvvz6st5uhv40tys",
+				Hash:       []byte{0x16, 0x53, 0x23, 0x70, 0x15, 0x89, 0xb7, 0x20, 0x14, 0xd5, 0xbd, 0xdc, 0xa8, 0xba, 0xcc, 0x60, 0xb5, 0x5, 0xd3, 0x97},
+				Balance: storage.Balance{
+					Id:    0,
+					Total: decimal.Zero,
+				},
+			},
+		},
+	}
+
+	msgExpected := storage.Message{
+		Id:       0,
+		Height:   blob.Height,
+		Time:     now,
+		Position: 0,
+		Type:     storageTypes.MsgPayForBlobs,
+		TxId:     0,
+		Data:     structs.Map(msgPayForBlob),
+		Namespace: []storage.Namespace{
+			{
+				Id:          0,
+				FirstHeight: blob.Height,
+				Version:     0,
+				NamespaceID: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 189, 44, 204, 197, 144, 206, 197, 121, 37, 22},
+				Size:        1,
+				PfbCount:    1,
+				Reserved:    false,
+			},
+		},
+		Addresses: addressesExpected,
+	}
+
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(1), dm.BlobsSize)
+	assert.Equal(t, msgExpected, dm.Msg)
+	assert.Equal(t, addressesExpected, dm.Addresses)
+}
